+++ vowpalwabbit/oaa.cc
@@ -107,15 +107,15 @@ namespace OAA {
         if (ld->label == INT_MAX)
           strcpy(label_buf," unknown");
         else
-          sprintf(label_buf,"%8i",ld->label);
+          sprintf(label_buf,"%8lu",(long unsigned int)ld->label);
 
-        fprintf(stderr, "%-10.6f %-10.6f %8ld %8.1f   %s %8i %8lu\n",
+        fprintf(stderr, "%-10.6f %-10.6f %8ld %8.1f   %s %8lu %8lu\n",
                 all.sd->sum_loss/all.sd->weighted_examples,
                 all.sd->sum_loss_since_last_dump / (all.sd->weighted_examples - all.sd->old_weighted_examples),
                 (long int)all.sd->example_number,
                 all.sd->weighted_examples,
                 label_buf,
-                *(prediction_t*)&ec->final_prediction,
+                (long unsigned int)*(prediction_t*)&ec->final_prediction,
                 (long unsigned int)ec->num_features);
      
         all.sd->sum_loss_since_last_dump = 0.0;
